+++ src/components/pageElements/CharacterFrame/CharacterFrame.tsx
@@ -21,7 +21,11 @@ export default function CharacterFrame({ name = 'Laura' }: Props) {
 
   return (
     <div className='portraitFrame' id='mainElement'>
-      <img src={MY_GRAVATOR} className={gravatarStyles} />
+      <img
+        src={MY_GRAVATOR}
+        className={gravatarStyles}
+        alt='User profile'
+      />
       <div className='name'>{name}</div>
       <div className='iconFrame'>
         <div className='iconResting' />
