+++ src/components/pageElements/Modal/GM/CreateQuest/index.ts
@@ -1,0 +1,3 @@
+import CreateQuest from './CreateQuest';
+
+export default CreateQuest;
