+++ src/components/pageElements/Modal/Modal.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 
+import classNames from 'classnames';
+
 import { Quest as QuestType, ModalTypes } from '../../../types';
 
 import GM from './GM';
@@ -18,13 +20,20 @@ export default function Modal({
   selectedQuest,
   setShownModal
 }: Props) {
+  const GMShown = showModal === ModalTypes.GM;
+  const questSelected = showModal === ModalTypes.Quest && selectedQuest;
+
+  const classes = classNames(
+    'ModalArea',
+    GMShown && 'questGMArea',
+    questSelected && 'questModalArea'
+  );
+
   const renderModal =
-    (showModal === ModalTypes.GM && (
-      <GM closeModal={() => setShownModal(undefined)} />
-    )) ||
-    (showModal === ModalTypes.Quest && selectedQuest && (
+    (GMShown && <GM closeModal={() => setShownModal(undefined)} />) ||
+    (selectedQuest && questSelected && (
       <Quest quest={selectedQuest} setShownModal={setShownModal} />
     ));
 
-  return <div className='ModalArea'>{renderModal}</div>;
+  return <div className={classes}>{renderModal}</div>;
 }
